Rename VSCodeAPI util to vscode
